+++ px_device_identity/__init__.py
@@ -3,8 +3,8 @@ from .device import Device
 from .sign import Sign
 from .log import Logger
 from .classes import (
-  DeviceClass,
-  DeviceRegistration,
-  RequestedOperation
+    DeviceClass,
+    DeviceRegistration,
+    RequestedOperation
 )
 from .config import DeviceConfig